requirements.txt
# ── FastAPI Backend ──────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# ── Database ────────────────────────────────────────────
sqlalchemy>=2.0.25
alembic>=1.13.0

# ── Scheduler ───────────────────────────────────────────
apscheduler>=3.10.4

# ── Existing Dependencies (from original app) ──────────
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
requests>=2.31.0
feedparser>=6.0.10
vaderSentiment>=3.3.2
scipy>=1.11.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytz>=2023.3
openai>=1.12.0

# ── Optional ────────────────────────────────────────────
# redis>=5.0.0  # Uncomment when using Redis instead of in-memory cache
